changed venue migrations lat/lng to FLOAT from DECIMAL

+++ backend/db/migrations/20231127205815-create-venue.js
@@ -35,10 +35,10 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       lng: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
